break down kverify to files
